Refactor components to separate files
