+++ .env
@@ -8,3 +8,6 @@ REACT_APP_API_HOSTNAME="https://localhost:8087"
 
 # Google Analytics' Universal ID
 ANALYTICS_ID="UA-208192118-1"
+
+# GeoLocation Rapid Api keu
+GEO_API_KEY="APIKEY"
