+++ src/components/Offers/New/CreateOfferForm.js
@@ -5,34 +5,24 @@ import {
     DialogContent,
     FormHelperText,
     Grid,
-    TextField,
     FormControl,
     Typography,
     Collapse,
     Button,
-    Slider,
-    Checkbox,
-    FormControlLabel,
-    MenuItem,
-    CircularProgress,
-    Tooltip,
 } from "@material-ui/core";
 import React, { useCallback, useContext, useEffect } from "react";
 import { Controller, useFieldArray, useForm, useWatch } from "react-hook-form";
 import { useMobile } from "../../../utils/media-queries";
 import CreateOfferSchema from "./CreateOfferSchema";
 import useCreateOfferStyles from "./createOfferStyles";
-import { CreateOfferConstants, defaultDates, parseRequestErrors } from "./CreateOfferUtils";
+import { defaultDates, parseRequestErrors } from "./CreateOfferUtils";
 import "./editor.css";
 import {
     KeyboardArrowDown,
     KeyboardArrowUp,
-    Info,
 } from "@material-ui/icons";
-import { KeyboardDatePicker } from "@material-ui/pickers";
 
-import JobOptions from "../../utils/offers/JobOptions";
-import { INITIAL_JOB_DURATION, JOB_MIN_DURATION, JOB_MAX_DURATION } from "../../../reducers/searchOffersReducer";
+import { INITIAL_JOB_DURATION } from "../../../reducers/searchOffersReducer";
 import MultiOptionAutocomplete from "../../utils/form/MultiOptionAutocomplete";
 import useFieldSelector from "../../utils/offers/useFieldSelector";
 import useTechSelector from "../../utils/offers/useTechSelector";
@@ -40,9 +30,19 @@ import { newOffer } from "../../../services/offerService";
 import useSession from "../../../hooks/useSession";
 import { Redirect } from "react-router-dom";
 import MultiOptionTextField from "../../utils/form/MultiOptionTextField";
-import TextEditor from "../../utils/TextEditor";
-import LocationPicker from "../../utils/LocationPicker";
 import ConnectedLoginAlert from "../../utils/LoginAlert";
+import TitleComponent from "./form-components/TitleComponent";
+import OwnerComponent from "./form-components/OwnerComponent";
+import LocationComponent from "./form-components/LocationComponent";
+import JobTypeComponent from "./form-components/JobTypeComponent";
+import JobStartDateComponent from "./form-components/JobStartDateComponent";
+import JobDurationComponent from "./form-components/JobDurationComponent";
+import VacanciesComponent from "./form-components/VacanciesComponent";
+import IsPaidComponent from "./form-components/IsPaidComponent";
+import PublicationDateComponent from "./form-components/PublicationDateComponent";
+import PublicationEndDateComponent from "./form-components/PublicationEndDateComponent";
+import IsHiddenComponent from "./form-components/IsHiddenComponent";
+import TextEditorComponent from "./form-components/TextEditorComponent";
 
 export const CreateOfferControllerContext = React.createContext();
 
@@ -61,28 +61,26 @@ export const CreateOfferController = () => {
     const companyUnfinishedRegistration = session.data?.company?.hasFinishedRegistration === false;
     const isLoggedIn = session.isLoggedIn;
 
-    // eslint-disable-next-line no-unused-vars
     const { handleSubmit, formState: { errors }, control, setValue, getValues, clearErrors } = useForm({
         mode: "all",
         resolver: yupResolver(CreateOfferSchema),
         reValidateMode: "onChange",
         defaultValues: {
             title: "",
-            publishDate: defaultDates.publishDate(),
-            publishEndDate: defaultDates.publishEndDate(),
+            publishDate: defaultDates.getPublishDate(),
+            publishEndDate: defaultDates.getPublishEndDate(),
             jobDuration: [INITIAL_JOB_DURATION, INITIAL_JOB_DURATION + 1],
             jobStartDate: null,
             description: "",
             descriptionText: "",
             contacts: [{ value: "" }],
             isPaid: "",
-            // https://stackoverflow.com/questions/37427508/react-changing-an-uncontrolled-input
             vacancies: "",
             jobType: "",
             fields: [],
             technologies: [],
             location: null,
-            requirements: [],
+            requirements: [{ value: "" }],
             isHidden: false,
             owner: "",
         },
@@ -95,10 +93,15 @@ export const CreateOfferController = () => {
         name: "requirements",
     });
 
+    const handleAppendRequirement = () => appendRequirement(DEFAULT_VALUE);
+
     const { fields: contacts, append: appendContact, remove: removeContact } = useFieldArray({
         control,
         name: "contacts",
     });
+
+    const handleAppendContact = () => appendContact(DEFAULT_VALUE);
+
 
     const [loading, setLoading] = React.useState(false);
     const [success, setSuccess] = React.useState(false);
@@ -122,6 +125,7 @@ export const CreateOfferController = () => {
                 contacts: data.contacts.map((val) => val.value),
                 requirements: data.requirements.map((val) => (val, val.value)),
                 isPaid: data.isPaid === "" ? undefined : data.isPaid,
+                jobStartDate: !data.jobStartDate ? undefined : data.jobStartDate,
                 owner: data.owner || company,
             })
                 .then((obj) => {
@@ -151,9 +155,9 @@ export const CreateOfferController = () => {
                 fields,
                 errors,
                 requestErrors,
-                appendContact: () => appendContact(DEFAULT_VALUE),
+                appendContact: handleAppendContact,
                 removeContact,
-                appendRequirement: () => appendRequirement(DEFAULT_VALUE),
+                appendRequirement: handleAppendRequirement,
                 removeRequirement,
                 getValues,
                 setValue,
@@ -211,12 +215,10 @@ const CreateOfferForm = () => {
     const Content = isMobile ? DialogContent : CardContent;
     const classes = useCreateOfferStyles(isMobile)();
 
-
     const FieldsSelectorProps = useFieldSelector(fields.fields, (fields) => setValue("fields", fields));
     const TechSelectorProps = useTechSelector(fields.technologies, (fields) => setValue("technologies", fields));
 
     return (
-
         success
             ? <Redirect to={`/offer/${newOfferId}`} push />
             :
@@ -237,121 +239,43 @@ const CreateOfferForm = () => {
                             >
                                 <Grid container>
                                     <Grid item xs={12} lg={isAdmin ? 6 : 12}>
-                                        <Controller
-                                            name="title"
-                                            render={(
-                                                { field: { onChange, onBlur, ref, name, value } },
-                                            ) => (
-                                                <TextField
-                                                    name={name}
-                                                    value={value}
-                                                    label="Offer Title"
-                                                    id="title"
-                                                    error={!!errors.title || !!requestErrors.title}
-                                                    inputRef={ref}
-                                                    onBlur={onBlur}
-                                                    onChange={onChange}
-                                                    helperText={
-                                                        `${value?.length}/${CreateOfferConstants.title.maxLength} 
-                                        ${errors.title?.message || requestErrors.title?.message || ""}`
-                                                    }
-                                                    variant="outlined"
-                                                    margin="dense"
-                                                    fullWidth
-                                                    disabled={disabled}
-                                                />)}
+                                        <TitleComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
                                             control={control}
                                         />
                                     </Grid>
 
                                     {isAdmin &&
                                     <Grid item xs={12} lg={6}>
-                                        <Controller
-                                            name="owner"
-                                            render={(
-                                                { field: { onChange, onBlur, ref, name, value } },
-                                            ) => (
-                                                <TextField
-                                                    name={name}
-                                                    value={value}
-                                                    label="Owner ID"
-                                                    id="owner"
-                                                    error={!!errors.owner || !!requestErrors.owner}
-                                                    inputRef={ref}
-                                                    onBlur={onBlur}
-                                                    onChange={onChange}
-                                                    helperText={
-                                                        `${errors.owner?.message || requestErrors.owner?.message || ""}`
-                                                    }
-                                                    variant="outlined"
-                                                    margin="dense"
-                                                    fullWidth
-                                                    disabled={disabled}
-                                                />)}
+                                        <OwnerComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
                                             control={control}
                                         />
                                     </Grid>}
 
                                     <Grid item xs={12} lg={6}>
                                         <FormControl fullWidth margin="dense">
-                                            <Controller
-                                                name="location"
-                                                render={(
-                                                    { field: { onChange, onBlur, name, value } },
-                                                ) => (
-                                                    <LocationPicker
-                                                        value={value}
-                                                        onChange={(_e, value) => onChange(value)}
-                                                        onBlur={onBlur}
-                                                        name={name}
-                                                        error={errors.location || requestErrors.location}
-                                                        disabled={disabled}
-                                                    />
-                                                )}
+                                            <LocationComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
                                                 control={control}
                                             />
-
                                         </FormControl>
 
                                     </Grid>
                                     <Grid item xs={12} lg={6}>
                                         <FormControl fullWidth margin="dense">
-
-                                            <Controller
-                                                name="jobType"
-                                                render={(
-                                                    { field: { onChange, onBlur, name, value } },
-                                                ) => (
-                                                    <TextField
-                                                        name={name}
-                                                        fullWidth
-                                                        id="job_type"
-                                                        select
-                                                        label="Job Type"
-                                                        value={value ? value : ""}
-                                                        onChange={onChange}
-                                                        onBlur={onBlur}
-                                                        variant="outlined"
-                                                        disabled={disabled}
-                                                        error={!!errors?.jobType || !!requestErrors.jobType}
-                                                        helperText={
-                                                            `${errors.jobType?.message || requestErrors.jobType?.message || ""}`
-                                                        }
-                                                    >
-                                                        {JobOptions.map(({ value, label }) => (
-                                                            <MenuItem
-                                                                key={value}
-                                                                value={value}
-                                                            >
-                                                                {label}
-                                                            </MenuItem>
-                                                        ))}
-                                                    </TextField>
-
-                                                )}
+                                            <JobTypeComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
                                                 control={control}
                                             />
-
                                         </FormControl>
                                     </Grid>
 
@@ -392,132 +316,40 @@ const CreateOfferForm = () => {
                                     </Grid>
                                     <Grid item xs={12} lg={6}>
                                         <FormControl>
-                                            <Controller
-                                                name="jobStartDate"
-                                                render={(
-                                                    { field: { onChange, onBlur, name, value } },
-                                                ) => (
-                                                    <KeyboardDatePicker
-                                                        margin="dense"
-                                                        value={value}
-                                                        label="Job Start Date"
-                                                        id="startDate-input"
-                                                        name={name}
-                                                        onChange={(_, value) => (onChange(value))}
-                                                        onBlur={onBlur}
-                                                        variant="inline"
-                                                        autoOk
-                                                        disabled={disabled}
-                                                        format="yyyy-MM-dd"
-                                                        minDate={Date.now()}
-                                                        error={!!errors?.jobStartDate || !!requestErrors.jobStartDate}
-                                                        helperText={
-                                                            `${errors.jobStartDate?.message ||
-                                                    requestErrors.jobStartDate?.message || " "}`
-                                                        }
-                                                    />)}
+                                            <JobStartDateComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
                                                 control={control}
                                             />
 
                                         </FormControl>
                                     </Grid>
                                     <Grid item xs={12} lg={6}>
-
-                                        <Controller
-                                            name="jobDuration"
-                                            render={(
-                                                { field: { onChange, onBlur, name, value } },
-                                            ) => (
-                                                <FormControl
-                                                    margin="normal"
-                                                    variant="outlined"
-                                                    fullWidth
-                                                >
-                                                    <Slider
-                                                        name={name}
-                                                        value={value}
-                                                        onChange={(_e, values) => onChange(values)}
-                                                        onBlur={onBlur}
-                                                        valueLabelDisplay="auto"
-                                                        aria-labelledby="range-slider"
-                                                        min={JOB_MIN_DURATION}
-                                                        max={JOB_MAX_DURATION}
-                                                        disabled={disabled}
-                                                    />
-
-                                                    <FormHelperText>
-                                                        {!disabled && `Job duration: ${value[0]} - ${value[1]} month(s)`}
-                                                    </FormHelperText>
-
-                                                </FormControl>)}
+                                        <JobDurationComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
                                             control={control}
                                         />
-
                                     </Grid>
 
                                     <Grid item xs={12} lg={6}>
                                         <FormControl>
-                                            <Controller
-                                                name="vacancies"
-                                                render={(
-                                                    { field: { onChange, onBlur, name, ref, value } },
-                                                ) => (
-                                                    <TextField
-                                                        name={name}
-                                                        value={value}
-                                                        label="Vacancies"
-                                                        id="vacancies"
-                                                        disabled={disabled}
-                                                        error={!!errors?.vacancies || !!requestErrors.vacancies}
-                                                        helperText={
-                                                            `${errors.vacancies?.message ||
-                                                                            requestErrors.vacancies?.message || ""}`
-                                                        }
-                                                        inputRef={ref}
-                                                        onChange={(_e) => {
-                                                            let value = _e.target.value.replace(/[^0-9]/g, "");
-                                                            value = value ? Number.parseInt(value, 10) : "";
-                                                            onChange(value);
-                                                        }}
-                                                        onBlur={onBlur}
-                                                    />)}
+                                            <VacanciesComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
                                                 control={control}
                                             />
                                         </FormControl>
                                     </Grid>
 
                                     <Grid item xs={12} lg={6}>
-                                        <Controller
-                                            name="isPaid"
-                                            render={(
-                                                { field: { onChange, onBlur, name, value } },
-                                            ) => (
-                                                <TextField
-                                                    name={name}
-                                                    fullWidth
-                                                    id="is-paid"
-                                                    select
-                                                    label="Compensation"
-                                                    value={value}
-                                                    onChange={onChange}
-                                                    onBlur={onBlur}
-                                                    variant="outlined"
-                                                    disabled={disabled}
-                                                    error={!!errors?.isPaid || !!requestErrors.isPaid}
-                                                    helperText={
-                                                        `${errors.isPaid?.message || requestErrors.isPaid?.message || " "}`
-                                                    }
-                                                >
-                                                    {PAID_OPTIONS.map(({ value, label }) => (
-                                                        <MenuItem
-                                                            key={value}
-                                                            value={value}
-                                                        >
-                                                            {label}
-                                                        </MenuItem>
-                                                    ))}
-                                                </TextField>
-                                            )}
+                                        <IsPaidComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
                                             control={control}
                                         />
                                     </Grid>
@@ -543,113 +375,37 @@ const CreateOfferForm = () => {
                                         >
                                             <Grid container>
                                                 <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
-
-                                                    <Controller
-                                                        name="publishDate"
-                                                        render={(
-                                                            { field: { onChange, onBlur, name, value } },
-                                                        ) => (
-                                                            <KeyboardDatePicker
-                                                                margin="dense"
-                                                                value={value}
-                                                                label="Publication Date"
-                                                                id="publishDate-input"
-                                                                name={name}
-                                                                disabled={disabled}
-                                                                onChange={(_, value) => onChange(value)}
-                                                                onBlur={onBlur}
-                                                                variant="inline"
-                                                                autoOk
-                                                                format="yyyy-MM-dd"
-                                                                minDate={Date.now()}
-                                                                error={!!errors?.publishDate || !!requestErrors?.publishDate }
-                                                                helperText={errors.publishDate?.message ||
-                                                    requestErrors.publishDate?.message || " "}
-                                                            />)}
-                                                        control={control}
-                                                    />
-
-                                                    <Tooltip
-                                                        title="The offer will only be visible after this date"
-                                                        placement="top"
-                                                    >
-                                                        <Info fontSize="small" />
-                                                    </Tooltip>
-                                                </Grid>
-                                                <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
-
-                                                    <Controller
-                                                        name="publishEndDate"
-                                                        render={(
-                                                            { field: { onChange, onBlur, name, value } },
-                                                        ) => (
-                                                            <KeyboardDatePicker
-                                                                margin="dense"
-                                                                value={value}
-                                                                label="Publication End Date"
-                                                                id="publishEndDate-input"
-                                                                name={name}
-                                                                disabled={disabled}
-                                                                onChange={(_, value) => {
-                                                                    const date = new Date(value);
-                                                                    date.setHours(23, 59, 59, 0);
-                                                                    onChange(date);
-                                                                }}
-                                                                onBlur={onBlur}
-                                                                variant="inline"
-                                                                autoOk
-                                                                format="yyyy-MM-dd"
-                                                                minDate={fields.publishDate}
-                                                                error={!!errors?.publishEndDate ||
-                                                                                 !!requestErrors.publishEndDate}
-                                                                helperText={errors.publishEndDate?.message ||
-                                                    requestErrors.publishEndDate?.message || " "}
-                                                            />)}
-                                                        control={control}
-                                                    />
-
-                                                    <Tooltip
-                                                        title="The offer will only be visible until this date"
-                                                        placement="top"
-                                                    >
-                                                        <Info fontSize="small" />
-                                                    </Tooltip>
-                                                </Grid>
-
-                                                <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
-                                                    <FormControlLabel
-                                                        label="Hide offer"
+                                                    <PublicationDateComponent
                                                         disabled={disabled}
-                                                        control={
-                                                            <Controller
-                                                                name="isHidden"
-                                                                render={(
-                                                                    { field: { onChange, onBlur, name, value } },
-                                                                ) => (
-                                                                    <Checkbox
-                                                                        checked={value}
-                                                                        onChange={onChange}
-                                                                        name={name}
-                                                                        onBlur={onBlur}
-                                                                        disabled={disabled}
-                                                                    />
-                                                                )}
-                                                                control={control}
-                                                            />
-                                                        }
+                                                        errors={errors}
+                                                        requestErrors={requestErrors}
+                                                        control={control}
                                                     />
-                                                    <Tooltip
-                                                        title="If the offer is hidden, users won't be able to see it"
-                                                        placement="top"
-                                                    >
-                                                        <Info fontSize="small" />
-                                                    </Tooltip>
+                                                </Grid>
+                                                <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
+                                                    <PublicationEndDateComponent
+                                                        fields={fields}
+                                                        disabled={disabled}
+                                                        errors={errors}
+                                                        requestErrors={requestErrors}
+                                                        control={control}
+                                                    />
+                                                </Grid>
+
+                                                <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
+                                                    <IsHiddenComponent
+                                                        disabled={disabled}
+                                                        errors={errors}
+                                                        requestErrors={requestErrors}
+                                                        control={control}
+                                                    />
                                                 </Grid>
                                             </Grid>
 
                                         </Collapse>
                                     </Grid>
                                 </Grid>
+
                                 <MultiOptionTextField
                                     values={contacts}
                                     label="Contacts"
@@ -678,31 +434,14 @@ const CreateOfferForm = () => {
                                     addEntryBtnTestId="requirements-selector"
                                 />
 
-                                <Controller
-                                    name="descriptionText"
-                                    render={(
-                                        { field: { onChange: onChangeDescriptionText } },
-                                    ) => (
-                                        <Controller
-                                            name="description"
-                                            render={(
-                                                { field: { onChange: onChangeDescription } },
-                                            ) => (
-                                                <TextEditor
-                                                    onChangeDescription={onChangeDescription}
-                                                    onChangeDescriptionText={onChangeDescriptionText}
-                                                    error={!!errors?.descriptionText || !!requestErrors?.descriptionText}
-                                                    content={fields.description}
-                                                    helperText={errors.descriptionText?.message ||
-                                        requestErrors.descriptionText?.message || " "}
-                                                    disabled={disabled}
-                                                />
-                                            )}
-                                            control={control}
-                                        />
-                                    )}
+                                <TextEditorComponent
+                                    fields={fields}
+                                    disabled={disabled}
+                                    errors={errors}
+                                    requestErrors={requestErrors}
                                     control={control}
                                 />
+
                                 <Grid item xs={12} lg={12} />
                                 {requestErrors.generalErrors ?
                                     requestErrors.generalErrors.map((error, idx) => (
@@ -722,20 +461,12 @@ const CreateOfferForm = () => {
                                 >
                 Submit
                                 </Button>
-                                {loading &&
-                                <CircularProgress
-                                    size={24}
-                                    className={classes.finishProgress}
-                                />
-                                }
                             </form>
                         </Grid>
                     </Grid>
                 </Content>
 
             </div>
-
-
     );
 };
