+++ src/components/Offers/New/form-components/PublicationEndDateComponent.js
@@ -1,0 +1,58 @@
+import React from "react";
+import { Controller } from "react-hook-form";
+import PropTypes from "prop-types";
+import { KeyboardDatePicker } from "@material-ui/pickers";
+import { Tooltip } from "@material-ui/core";
+import { Info } from "@material-ui/icons";
+
+
+const PublicationEndDateComponent = ({ fields, disabled, errors, requestErrors, control }) => (
+    <>
+        <Controller
+            name="publishEndDate"
+            render={(
+                { field: { onChange, onBlur, name, value } },
+            ) => (
+                <KeyboardDatePicker
+                    margin="dense"
+                    value={value}
+                    label="Publication End Date"
+                    id="publishEndDate-input"
+                    name={name}
+                    disabled={disabled}
+                    onChange={(_, value) => {
+                        const date = new Date(value);
+                        date.setHours(23, 59, 59, 0);
+                        onChange(date);
+                    }}
+                    onBlur={onBlur}
+                    variant="inline"
+                    autoOk
+                    format="yyyy-MM-dd"
+                    minDate={fields.publishDate}
+                    error={!!errors?.publishEndDate ||
+                                                                                 !!requestErrors.publishEndDate}
+                    helperText={errors.publishEndDate?.message ||
+                                                    requestErrors.publishEndDate?.message || " "}
+                />)}
+            control={control}
+        />
+
+        <Tooltip
+            title="The offer will only be visible until this date"
+            placement="top"
+        >
+            <Info fontSize="small" />
+        </Tooltip>
+    </>
+);
+
+PublicationEndDateComponent.propTypes = {
+    fields: PropTypes.object.isRequired,
+    disabled: PropTypes.bool,
+    errors: PropTypes.object,
+    requestErrors: PropTypes.object,
+    control: PropTypes.object.isRequired,
+};
+
+export default PublicationEndDateComponent;
