+++ src/components/utils/TextEditor.js
@@ -14,10 +14,13 @@ const useStyles = makeStyles((theme) => ({
     editorToolbar: {
         marginBottom: theme.spacing(1),
         display: "flex",
+        flexWrap: "wrap",
     },
     headingBtn: {
-        width: "36px",
-        height: "36px",
+        height: theme.spacing(4.5),
+    },
+    section: {
+        fontSize: "5em",
     },
 }));
 
@@ -37,9 +40,7 @@ const EditorToolbar = ({ editor, disabled }) => {
             if (editor.isActive("underline")) state.push("underline");
             if (editor.isActive("bulletList")) state.push("bulletList");
             if (editor.isActive("orderedList")) state.push("orderedList");
-            for (let level = 1; level < 6; level++) {
-                if (editor.isActive("heading", { level })) state.push(`heading${level}`);
-            }
+            if (editor.isActive("heading", { level: 1 })) state.push("heading1");
             setFormats(state);
         };
 
@@ -54,7 +55,7 @@ const EditorToolbar = ({ editor, disabled }) => {
     return (
 
         <div className={classes.editorToolbar}>
-            <Box mr={1} display="inline">
+            <Box mr={2} mb={1} display="inline">
                 <ToggleButtonGroup size="small" value={formats} aria-label="text formatting">
                     <ToggleButton
                         value="bold"
@@ -82,7 +83,7 @@ const EditorToolbar = ({ editor, disabled }) => {
                     </ToggleButton>
                 </ToggleButtonGroup>
             </Box>
-            <Box ml={1} mr={1} display="inline">
+            <Box mr={2} mb={1} display="inline">
                 <ToggleButtonGroup size="small" value={formats} aria-label="text lists">
                     <ToggleButton
                         value="bulletList"
@@ -102,7 +103,7 @@ const EditorToolbar = ({ editor, disabled }) => {
                     </ToggleButton>
                 </ToggleButtonGroup>
             </Box>
-            <Box ml={1} display="inline">
+            <Box display="inline">
                 <ToggleButtonGroup size="small" value={formats} aria-label="text lists">
                     <ToggleButton
                         className={classes.headingBtn}
@@ -111,47 +112,7 @@ const EditorToolbar = ({ editor, disabled }) => {
                         onClick={() => editor.chain().focus().toggleHeading({ level: 1 }).run()}
                         disabled={disabled}
                     >
-                        <Typography>h1</Typography>
-                    </ToggleButton>
-
-                    <ToggleButton
-                        className={classes.headingBtn}
-                        value="heading2"
-                        aria-label="heading2"
-                        onClick={() => editor.chain().focus().toggleHeading({ level: 2 }).run()}
-                        disabled={disabled}
-                    >
-                        <Typography>h2</Typography>
-                    </ToggleButton>
-
-                    <ToggleButton
-                        className={classes.headingBtn}
-                        value="heading3"
-                        aria-label="heading3"
-                        onClick={() => editor.chain().focus().toggleHeading({ level: 3 }).run()}
-                        disabled={disabled}
-                    >
-                        <Typography>h3</Typography>
-                    </ToggleButton>
-
-                    <ToggleButton
-                        className={classes.headingBtn}
-                        value="heading4"
-                        aria-label="heading4"
-                        onClick={() => editor.chain().focus().toggleHeading({ level: 4 }).run()}
-                        disabled={disabled}
-                    >
-                        <Typography>h4</Typography>
-                    </ToggleButton>
-
-                    <ToggleButton
-                        className={classes.headingBtn}
-                        value="heading5"
-                        aria-label="heading5"
-                        onClick={() => editor.chain().focus().toggleHeading({ level: 5 }).run()}
-                        disabled={disabled}
-                    >
-                        <Typography>h5</Typography>
+                        <Typography>section</Typography>
                     </ToggleButton>
                 </ToggleButtonGroup>
             </Box>
@@ -165,7 +126,6 @@ EditorToolbar.propTypes =  {
 };
 
 const TextEditor = ({ content, onChangeDescription, onChangeDescriptionText, error, helperText: additionalHelperText, disabled }) => {
-
     const editor = useEditor({
         extensions: [
             StarterKit,
@@ -174,13 +134,12 @@ const TextEditor = ({ content, onChangeDescription, onChangeDescriptionText, err
                 placeholder: "Write a description for this offer. You can specify goals, project and daily work details, etc.",
             }),
             Heading.configure({
-                levels: [1, 2, 3, 4, 5],
+                levels: [1],
             }),
         ],
         content,
         editable: true,
         editorProps: {
-
             attributes: {
                 class: "editor", // Cannot use makeStyles with this since it won't update the class name on re-render :(
             },
